refactor: 优化代码结构和功能
- 移除 CategorySnapshot 相关的实体类和接口
- 更新 about.html 中的版本号和构建日期
- 修改 github.html 的语言属性
- 优化 I18nServiceImpl 中的字符串检查逻辑
- 更新用户欢迎信息的显示顺序
- 修正 orders.html 中的 jQuery 文件路径

+++ src/main/java/com/jiang/mall/controller/ProductController.java
@@ -167,7 +167,7 @@ public class ProductController {
     @PostMapping("/add")
     public ResponseResult insertProduct(@RequestParam("code") String code,
                                         @RequestParam("title") String title,
-                                        @RequestParam("categoryId") Integer categoryId,
+                                        @RequestParam("categoryId") Long categoryId,
                                         @RequestParam("img") String img,
                                         @RequestParam("price") Double price,
                                         @RequestParam("stocks") Integer stocks,
@@ -231,7 +231,7 @@ public class ProductController {
     public ResponseResult updateProduct(@RequestParam("id") Long id,
                                         @RequestParam("code") String code,
                                         @RequestParam("title") String title,
-                                        @RequestParam("categoryId") Integer categoryId,
+                                        @RequestParam("categoryId") Long categoryId,
                                         @RequestParam("img") String img,
                                         @RequestParam("price") Double price,
                                         @RequestParam("stocks") Integer stocks,
